Add revision to entity files
This gives the user the possibility to make entities "reproducible",
because an arbitrary revision can be specified. Not only a branch, but
also a tag or just a commit ID. It doesn't matter as long as 'git
fetch' understands it.
The revision can also be empty in which case nothing will be checked
out after cloning, it will stay on the default branch.